Ignore invalid regions instead of wasting time searching for them
Note to addon authors: DO NOT use GetCVar("portal") and pass that to LRI. If you want to get info about a realm in the player's current region, just pass the realm name without specifying a region. The "portal" CVar is unreliable and does not match the region codes used by LRI anyway. RTFM!