Remove realm entries for nameless realms
These are used to host connected realm groups. They are not included
in the realm info from the Battle.net API, but are included (as the
connected group ID) from the connected realm info API. To support
region detection, we now add these to the realm info table dynamically
when unpacking the connected realm info.